refactor: add `ImportMetaContextDependencyParserPlugin`
