+++ crates/rspack_plugin_javascript/src/parser_plugin/mod.rs
@@ -9,6 +9,7 @@ mod drive;
 mod exports_info_api_plugin;
 mod harmony_detection_parser_plugin;
 mod harmony_top_level_this_plugin;
+mod import_meta_context_dependency_parser_plugin;
 mod node_stuff_plugin;
 mod provide;
 mod require_context_dependency_parser_plugin;
@@ -29,6 +30,7 @@ pub(crate) use self::drive::JavaScriptParserPluginDrive;
 pub(crate) use self::exports_info_api_plugin::ExportsInfoApiPlugin;
 pub(crate) use self::harmony_detection_parser_plugin::HarmonDetectionParserPlugin;
 pub(crate) use self::harmony_top_level_this_plugin::HarmonyTopLevelThisParserPlugin;
+pub(crate) use self::import_meta_context_dependency_parser_plugin::ImportMetaContextDependencyParserPlugin;
 pub(crate) use self::node_stuff_plugin::NodeStuffPlugin;
 pub(crate) use self::provide::ProviderPlugin;
 pub(crate) use self::r#const::{is_logic_op, ConstPlugin};
