+++ crates/rspack_plugin_javascript/src/visitors/dependency/import_scanner.rs
@@ -1,13 +1,12 @@
 use std::sync::Arc;
 
 use rspack_core::{
-  clean_regexp_in_context_module, context_reg_exp, AsyncDependenciesBlock, DependencyLocation,
-  DynamicImportMode, ErrorSpan, GroupOptions, JavascriptParserOptions, ModuleIdentifier,
+  context_reg_exp, AsyncDependenciesBlock, DependencyLocation, DynamicImportMode, ErrorSpan,
+  GroupOptions, JavascriptParserOptions, ModuleIdentifier,
 };
 use rspack_core::{BoxDependency, BuildMeta, ChunkGroupOptions, ContextMode};
 use rspack_core::{ContextNameSpaceObject, ContextOptions, DependencyCategory, SpanExt};
 use rspack_error::miette::Diagnostic;
-use rspack_regex::{regexp_as_str, RspackRegex};
 use rustc_hash::FxHashSet;
 use swc_core::common::comments::Comments;
 use swc_core::common::{SourceFile, Spanned};
@@ -16,11 +15,10 @@ use swc_core::ecma::atoms::Atom;
 use swc_core::ecma::visit::{noop_visit_type, Visit, VisitWith};
 
 use super::context_helper::scanner_context_module;
-use super::{is_import_meta_context_call, parse_order_string};
+use super::parse_order_string;
+use crate::dependency::ImportEagerDependency;
 use crate::dependency::{ImportContextDependency, ImportDependency};
-use crate::dependency::{ImportEagerDependency, ImportMetaContextDependency};
 use crate::no_visit_ignored_stmt;
-use crate::utils::{get_bool_by_obj_prop, get_literal_str_by_obj_prop, get_regex_by_obj_prop};
 use crate::webpack_comment::try_extract_webpack_magic_comment;
 
 pub struct ImportScanner<'a> {
@@ -33,83 +31,6 @@ pub struct ImportScanner<'a> {
   pub options: Option<&'a JavascriptParserOptions>,
   pub warning_diagnostics: &'a mut Vec<Box<dyn Diagnostic + Send + Sync>>,
   pub ignored: &'a mut FxHashSet<DependencyLocation>,
-}
-
-fn create_import_meta_context_dependency(node: &CallExpr) -> Option<ImportMetaContextDependency> {
-  assert!(node.callee.is_expr());
-  let dyn_imported = node.args.first()?;
-  if dyn_imported.spread.is_some() {
-    return None;
-  }
-  let context = dyn_imported
-    .expr
-    .as_lit()
-    .and_then(|lit| {
-      if let Lit::Str(str) = lit {
-        return Some(str.value.to_string());
-      }
-      None
-    })
-    // TODO: should've used expression evaluation to handle cases like `abc${"efg"}`, etc.
-    .or_else(|| {
-      if let Some(tpl) = dyn_imported.expr.as_tpl()
-        && tpl.exprs.is_empty()
-        && tpl.quasis.len() == 1
-        && let Some(el) = tpl.quasis.first()
-      {
-        return Some(el.raw.to_string());
-      }
-      None
-    })?;
-  let reg = r"^\.\/.*$";
-  let reg_str = reg.to_string();
-  let context_options = if let Some(obj) = node.args.get(1).and_then(|arg| arg.expr.as_object()) {
-    let regexp = get_regex_by_obj_prop(obj, "regExp")
-      .map(|regexp| RspackRegex::try_from(regexp).expect("reg failed"))
-      .unwrap_or(RspackRegex::new(reg).expect("reg failed"));
-    // let include = get_regex_by_obj_prop(obj, "include")
-    //   .map(|regexp| RspackRegex::try_from(regexp).expect("reg failed"));
-    // let exclude = get_regex_by_obj_prop(obj, "include")
-    //   .map(|regexp| RspackRegex::try_from(regexp).expect("reg failed"));
-    let mode = get_literal_str_by_obj_prop(obj, "mode")
-      .map(|s| s.value.to_string().as_str().into())
-      .unwrap_or(ContextMode::Sync);
-    let recursive = get_bool_by_obj_prop(obj, "recursive")
-      .map(|bool| bool.value)
-      .unwrap_or(true);
-    let reg_str = regexp_as_str(&regexp).to_string();
-    ContextOptions {
-      chunk_name: None,
-      reg_exp: clean_regexp_in_context_module(regexp),
-      reg_str,
-      include: None,
-      exclude: None,
-      recursive,
-      category: DependencyCategory::Esm,
-      request: context,
-      namespace_object: ContextNameSpaceObject::Unset,
-      mode,
-    }
-  } else {
-    ContextOptions {
-      chunk_name: None,
-      recursive: true,
-      mode: ContextMode::Sync,
-      include: None,
-      exclude: None,
-      reg_exp: context_reg_exp(reg, ""),
-      reg_str,
-      category: DependencyCategory::Esm,
-      request: context,
-      namespace_object: ContextNameSpaceObject::Unset,
-    }
-  };
-  Some(ImportMetaContextDependency::new(
-    node.span.real_lo(),
-    node.span.real_hi(),
-    context_options,
-    Some(node.span.into()),
-  ))
 }
 
 impl<'a> ImportScanner<'a> {
@@ -144,13 +65,6 @@ impl Visit for ImportScanner<'_> {
   no_visit_ignored_stmt!();
 
   fn visit_call_expr(&mut self, node: &CallExpr) {
-    if !node.args.is_empty()
-      && is_import_meta_context_call(node)
-      && let Some(dep) = create_import_meta_context_dependency(node)
-    {
-      self.dependencies.push(Box::new(dep));
-      return;
-    }
     let Callee::Import(import_call) = &node.callee else {
       node.visit_children_with(self);
       return;
